+++ lib/legacy.dart
@@ -1,0 +1,1 @@
+export 'src/models/old/models.dart';