+++ lib/src/models/models.dart
@@ -1,10 +1,9 @@
+export 'alergen.dart';
 export 'feature_support.dart';
+export 'jidelnicek.dart';
+export 'jidlo.dart';
+export 'prihlasovaci_udaje.dart';
+export 'slozeni_jidla.dart';
+export 'stav_jidla.dart';
+export 'stav_uctu.dart';
 export 'uzivatelske_udaje.dart';
-
-export 'old/alergen/alergen.dart';
-export 'old/jidlo_kategorizovano/jidlo_kategorizovano.dart';
-export 'old/login_data/login_data.dart';
-export 'old/burza/burza.dart';
-export 'old/jidelnicek/jidelnicek.dart';
-export 'old/jidlo/jidlo.dart';
-export 'old/uzivatel/uzivatel.dart';