+++ lib/src/version_specific/canteen_2_18_19.dart
@@ -25,11 +25,13 @@ SOFTWARE.
   testováno na webu http://obedy.zs-mat5.cz/
 */
 
-import 'package:html/dom.dart' as dom;
-import 'package:html/parser.dart' as parser;
-import 'package:http/http.dart' as http;
+import 'package:icanteenlib/legacy.dart' as legacy;
 import 'package:icanteenlib/canteenlib.dart';
 import 'package:icanteenlib/src/utils/string_utils.dart';
+
+import 'package:http/http.dart' as http;
+import 'package:html/dom.dart' as dom;
+import 'package:html/parser.dart' as parser;
 
 /// Reprezentuje kantýnu verze 2.18.19
 ///
@@ -82,7 +84,7 @@ class Canteen2v18v19 extends Canteen {
 
   /// Vrátí informace o uživateli ve formě instance [Uzivatel]
   @override
-  Future<Uzivatel> ziskejUzivatele() async {
+  Future<legacy.Uzivatel> ziskejUzivatele() async {
     if (!prihlasen) return Future.error(CanteenLibExceptions.jePotrebaSePrihlasit);
     String res;
     try {
@@ -98,7 +100,7 @@ class Canteen2v18v19 extends Canteen {
     String kredit = kreditElement?.text ?? '0.0';
     kredit = kredit.replaceAll(RegExp(r'[^0-9,.-]'), '').replaceAll(',', '.');
 
-    return Uzivatel(
+    return legacy.Uzivatel(
       uzivatelskeJmeno: _username,
       jmeno: userData[UzivatelskeUdajeKeys.jmeno],
       prijmeni: userData[UzivatelskeUdajeKeys.prijmeni],
@@ -207,7 +209,7 @@ class Canteen2v18v19 extends Canteen {
   }
 
   @override
-  Future<List<Jidelnicek>> verejnyJidelnicek() async {
+  Future<List<legacy.Jidelnicek>> verejnyJidelnicek() async {
     String res;
     try {
       res = await _getRequest("/");
@@ -219,7 +221,7 @@ class Canteen2v18v19 extends Canteen {
       dotAll: true,
     ).allMatches(res).toList();
 
-    List<Jidelnicek> jidelnicek = [];
+    List<legacy.Jidelnicek> jidelnicek = [];
 
     for (var t in reg) {
       // projedeme každý den individuálně
@@ -236,7 +238,7 @@ class Canteen2v18v19 extends Canteen {
         ).allMatches(j).toList(); // získáme jednotlivá jídla pro den / VERZE 2.10
       }
 
-      List<Jidlo> jidla = [];
+      List<legacy.Jidlo> jidla = [];
 
       for (var jidloNaDen in jidlaDenne) {
         // projedeme vsechna jidla
@@ -252,15 +254,17 @@ class Canteen2v18v19 extends Canteen {
           dotAll: true,
         ).firstMatch(s)!.group(1).toString(); // Jídlo
 
-        jidla.add(Jidlo(nazev: hlavni, objednano: false, varianta: vydejna!.group(0).toString(), lzeObjednat: false, den: den, naBurze: false));
+        jidla.add(
+          legacy.Jidlo(nazev: hlavni, objednano: false, varianta: vydejna!.group(0).toString(), lzeObjednat: false, den: den, naBurze: false),
+        );
       }
-      jidelnicek.add(Jidelnicek(den, jidla));
+      jidelnicek.add(legacy.Jidelnicek(den, jidla));
     }
     return jidelnicek;
   }
 
   @override
-  Future<List<Jidelnicek>> jidelnicekMesic() async {
+  Future<List<legacy.Jidelnicek>> jidelnicekMesic() async {
     if (!prihlasen) {
       return Future.error(CanteenLibExceptions.jePotrebaSePrihlasit);
     }
@@ -271,12 +275,12 @@ class Canteen2v18v19 extends Canteen {
       return Future.error(e);
     }
 
-    var jidla = <Jidlo>[];
+    var jidla = <legacy.Jidlo>[];
     var jidelnicek = RegExp(r'<div class="jidelnicekItemWrapper">([\s\S]+?)</div>\s*</div>', dotAll: true).allMatches(res).toList();
     for (var obed in jidelnicek) {
       jidla.add(_parsePrihlasenyJidlo(obed));
     }
-    Map<DateTime, List<Jidlo>> jidlaMap = {};
+    Map<DateTime, List<legacy.Jidlo>> jidlaMap = {};
     for (var jidlo in jidla) {
       if (jidlaMap.containsKey(jidlo.den)) {
         jidlaMap[jidlo.den]!.add(jidlo);
@@ -284,14 +288,14 @@ class Canteen2v18v19 extends Canteen {
         jidlaMap[jidlo.den] = [jidlo];
       }
     }
-    List<Jidelnicek> jidelnicekList = [];
+    List<legacy.Jidelnicek> jidelnicekList = [];
     for (var jidelnicek in jidlaMap.values) {
-      jidelnicekList.add(Jidelnicek(jidelnicek[0].den, jidelnicek));
+      jidelnicekList.add(legacy.Jidelnicek(jidelnicek[0].den, jidelnicek));
     }
     return jidelnicekList;
   }
 
-  Jidlo _parsePrihlasenyJidlo(RegExpMatch obed) {
+  legacy.Jidlo _parsePrihlasenyJidlo(RegExpMatch obed) {
     // formátování do třídy
     var obedFormated = obed.group(0).toString().replaceAll(RegExp(r'(   )+|([^>a-z]\n)'), '');
     var objednano = obedFormated.contains("Máte objednáno");
@@ -308,11 +312,11 @@ class Canteen2v18v19 extends Canteen {
       dotAll: true,
     ).firstMatch(obedFormated)!.group(1).toString().replaceAll(' ,', ",").replaceAll(" <br>", "").replaceAll("\n", "");
     var alergenyList = RegExp(r'(<span\s*title=.*?<\/span>)').allMatches(jidlaProDen).toList();
-    var alergeny = alergenyList.map<Alergen>((e) {
+    var alergeny = alergenyList.map<legacy.Alergen>((e) {
       var jmeno = RegExp(r'<b>(.+?)<\/b>').firstMatch(e.group(1).toString())!.group(1);
       var popis = RegExp(r'<\/b> - (.+)').firstMatch(e.group(1).toString())?.group(1);
       var kod = RegExp(r'class="textGrey">(\d+?),?\s?').firstMatch(e.group(1).toString())?.group(1);
-      return Alergen(nazev: jmeno!, kod: kod == null ? null : int.parse(kod), popis: popis);
+      return legacy.Alergen(nazev: jmeno!, kod: kod == null ? null : int.parse(kod), popis: popis);
     }).toList();
 
     var vydejna = RegExp(r'(?<=<span class="smallBoldTitle button-link-align">).+?(?=<)').firstMatch(obedFormated)!.group(0).toString();
@@ -334,7 +338,7 @@ class Canteen2v18v19 extends Canteen {
     }
     var jidloJmeno = jidlaProDen.split('<sub>')[0];
     jidloJmeno = cleanString(jidloJmeno);
-    return Jidlo(
+    return legacy.Jidlo(
       nazev: jidloJmeno.replaceAll(r' (?=[^a-zA-ZěščřžýáíéĚŠČŘŽÝÁÍÉŤŇťň])', ''),
       objednano: objednano,
       varianta: vydejna,
@@ -351,7 +355,7 @@ class Canteen2v18v19 extends Canteen {
   }
 
   @override
-  Future<Jidelnicek> jidelnicekDen({DateTime? den}) async {
+  Future<legacy.Jidelnicek> jidelnicekDen({DateTime? den}) async {
     if (!prihlasen) {
       return Future.error(CanteenLibExceptions.jePotrebaSePrihlasit);
     }
@@ -367,17 +371,17 @@ class Canteen2v18v19 extends Canteen {
       return Future.error(e);
     }
 
-    var jidla = <Jidlo>[];
+    var jidla = <legacy.Jidlo>[];
     var jidelnicek = RegExp(r'(?<=<div class="jidWrapLeft">).+?((fa-clock)|(fa-ban))', dotAll: true).allMatches(res).toList();
     for (var obed in jidelnicek) {
       jidla.add(_parsePrihlasenyJidlo(obed));
     }
 
-    return Jidelnicek(den, jidla);
+    return legacy.Jidelnicek(den, jidla);
   }
 
   @override
-  Future<Jidelnicek> objednat(Jidlo jidlo) async {
+  Future<legacy.Jidelnicek> objednat(legacy.Jidlo jidlo) async {
     if (!prihlasen) {
       return Future.error(CanteenLibExceptions.jePotrebaSePrihlasit);
     }
@@ -397,7 +401,7 @@ class Canteen2v18v19 extends Canteen {
   }
 
   @override
-  Future<Jidelnicek> doBurzy(Jidlo jidlo, {int amount = 1}) async {
+  Future<legacy.Jidelnicek> doBurzy(legacy.Jidlo jidlo, {int amount = 1}) async {
     if (!prihlasen) {
       return Future.error(CanteenLibExceptions.jePotrebaSePrihlasit);
     }
@@ -420,9 +424,9 @@ class Canteen2v18v19 extends Canteen {
   }
 
   @override
-  Future<List<Burza>> ziskatBurzu() async {
+  Future<List<legacy.Burza>> ziskatBurzu() async {
     if (!prihlasen) return Future.error(CanteenLibExceptions.jePotrebaSePrihlasit);
-    List<Burza> burza = [];
+    List<legacy.Burza> burza = [];
 
     String res;
     try {
@@ -451,7 +455,7 @@ class Canteen2v18v19 extends Canteen {
         var pocet = int.parse(data[4].group(0)!.replaceAll(" ks", ""));
         var url = RegExp(r"(?<=')db.+?(?=')").firstMatch(bu)!.group(0)!.replaceAll("&amp;", "&");
 
-        var jidlo = Burza(den: datum, varianta: varianta, nazev: nazev, pocet: pocet, url: url);
+        var jidlo = legacy.Burza(den: datum, varianta: varianta, nazev: nazev, pocet: pocet, url: url);
         burza.add(jidlo);
       }
     }
@@ -459,7 +463,7 @@ class Canteen2v18v19 extends Canteen {
   }
 
   @override
-  Future<Jidelnicek> objednatZBurzy(Burza b) async {
+  Future<legacy.Jidelnicek> objednatZBurzy(legacy.Burza b) async {
     if (!prihlasen) return Future.error(CanteenLibExceptions.jePotrebaSePrihlasit);
     try {
       await _getRequest("/faces/secured/${b.url!}");